+++ magik-checks/src/main/java/org/stevenlooman/sw/magik/checks/MethodDocCheck.java
@@ -66,6 +66,27 @@ public class MethodDocCheck extends MagikCheck {
         addIssue(message, node);
       }
     }
+
+    if (node.getTokenValue().equals("_iter")) {
+      // Require loopbody if it is an iterator method.
+      if (docParser.getSection("loopbody") == null) {
+        String message = String.format(MESSAGE, "loopbody");
+        addIssue(message, node);
+      } else {
+        // Match loopbody arguments with sub-section of loopbody.
+        int loopParameterCount = getLoopbodyParameterCount(node);
+        if (loopParameterCount != docParser.getLoopParameters().size()) {
+          String message = String.format(MESSAGE, "Loopbody parameter " + loopParameterCount);
+          addIssue(message, node);
+        }
+      }
+    } else {
+      // Disallow loopbody if it is an iterator method.
+      if (docParser.getSection("loopbody") != null) {
+        String message = String.format(MESSAGE, "loopbody");
+        addIssue(message, node);
+      }
+    }
   }
 
   private List<String> getMethodParameters(AstNode node) {
@@ -100,11 +121,30 @@ public class MethodDocCheck extends MagikCheck {
     AstNode assignmentParameterNode = node.getFirstChild(MagikGrammar.ASSIGNMENT_PARAMETER);
     if (assignmentParameterNode != null) {
       AstNode identifierNode = assignmentParameterNode.getFirstChild(MagikGrammar.IDENTIFIER);
-      String name = identifierNode.getTokenValue();
+      String name = identifierNode.getTokenValue().toLowerCase();
       parameters.add(name);
     }
 
     return parameters;
+  }
+
+  private int getLoopbodyParameterCount(AstNode node) {
+    List<AstNode> loopbodyNodes = node.getDescendants(MagikGrammar.LOOPBODY);
+
+    int max = 0;
+    for (AstNode loopbodyNode : loopbodyNodes) {
+      // Ensure part part of iter proc.
+      if (loopbodyNode.getFirstAncestor(MagikGrammar.PROC_DEFINITION) != null) {
+        continue;
+      }
+
+      AstNode expressionsNode = loopbodyNode.getFirstChild(MagikGrammar.EXPRESSIONS);
+      if (expressionsNode == null) {
+        continue;
+      }
+      max = Math.max(expressionsNode.getChildren(MagikGrammar.EXPRESSION).size(), max);
+    }
+    return max;
   }
 
 }