+++ CHANGES.md
@@ -19,6 +19,7 @@ Changes
 - Make parsing/checking files parallel in magik-lint
 - Check Formatting now requires empty line after TRANSMIT
 - Fix scoping for variables defined at \_for loops
+- MethodDocCheck now needs Loopbody-section for \_iter methods
 
 
 0.3.2 (2019-10-29)